Rev 1.1 of ESP32 Switchboard
I think this rev is basically ready to go. I could see someone using
this for real. I still have a lot of software features I wanna add, but
the hardware is basically complete.

I may come back and add a built-in serial port, as the little breakout
can be flimsy, but it seems that 3v3 rs232 stuff means SMD components,
and I've been trying to avoid that to make this easier to DIY.

board:
  Converted footprint to Kicad non-legacy
  Fixed footprint to be HiLetGo ESP-E32, which is narrower than NodeMCU
  Fixed up references on schematic to be in order
  Flipped mistaken footprint on RS232 UART breakout

code:
  Added support for all 12 inputs
  IRQ-based on rising/falling edges
  Added 0.1s timer-based timeout for input switch after input off
    Fixes some bouncing in the signal I didn't expect when powering off