Version 2.2 of extron input (move sense to top)
This revision moves the sense circuit to the top of the board, which
makes the audio plug a lot easier to get at, and you can throw your
esp32-switcher on the top of your extron.

Also cleaned up rescue files for eeschema. Come on, eeschema. Get your
shit together.